requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langchain-core>=0.3.3
langchain-ollama>=0.2.0
langgraph>=0.2.38
typing-extensions>=4.12.2
python-dotenv>=1.0.1
